Implement the behavior for disabled controls.
There was no mechanism in place, at the central level of UI inputt handling, for dealing with disabled controls. This has now been added. The 'IsDisabled' method accounts also for the status of the parent, therefore disabling a control will put all the children in this state, even though the children are not disabled intrinsically.